Fix ADFS validation
ADFS does not respond to HEAD requests, so using CURLOPT_NOBODY results
in validation always timing out. In order to emulate the effect of
NOBODY with a GET request, we use CURLOPT_RETURNTRANSFER to capture the
output of curl_exec() [rather than returning to browser] and then check
for the existance of a false return code to emulate the original return
codes with CURLOPT_NOBODY set. This effectively discards the body. It's
a bit wasteful bandwidth-wise, but should give a better result.